Rename job for consistency
